Make intro accessible to new comers
Currently the readme assumes familiarity with (an unlinked) previous non-Go iteration.

This commit makes it a bit more accessible to new-comers, such as myself. (Copy taken from the hosted instance linked.)

It's still not clear to me whether Plaxt syncs in both directions, or Plex -> Trakt only? It reads like the latter, but I read a prominent Plex community developer claiming on Reddit that Trakt -> Plex without a plugin was a solved problem, and in Plaxt lay the solution. I think the readme should make it clearer either way, but I'm not presently in a position to do that.